requirements.txt
requests>=2.31
pyyaml>=6.0
psycopg2-binary>=2.9
